Change threas per block in collatz, fix batch_size type and fix stack overflows (#94)
* Change threads per block in collatz

* Fix batch_size

* Change batch_size from int to ulong, fix stack overflow

+++ cudampilib/apps/collatz/app-streams-collatz.c
@@ -30,7 +30,7 @@ long long VECTORSIZE;
 double *vectora;
 double *vectorc;
 
-int batchsize;
+unsigned long batchsize;
 
 long long globalcounter = 0;
