+++ cudampilib/apps/patternsearch/appkernelpatternsearch.cu
@@ -45,7 +45,7 @@ __global__ void appkernel(void *devPtr) {
   }
 }
 
-extern "C" void launchkernelinstream(void *devPtr, int batchSize, cudaStream_t stream) {
+extern "C" void launchkernelinstream(void *devPtr, unsigned long batchSize, cudaStream_t stream) {
 
   dim3 blocksingrid(PATTERNSEARCH_BLOCKS_IN_GRID);
   dim3 threadsinblock(batchSize / PATTERNSEARCH_BLOCKS_IN_GRID);
@@ -58,4 +58,4 @@ extern "C" void launchkernelinstream(void *devPtr, int batchSize, cudaStream_t s
   }
 }
 
-extern "C" void launchkernel(void *devPtr, int batchSize) { launchkernelinstream(devPtr, batchSize, 0); }
+extern "C" void launchkernel(void *devPtr, unsigned long batchSize) { launchkernelinstream(devPtr, batchSize, 0); }
