+++ cudampilib/apps/vecmaxdiv/app-streams-vecmaxdiv.c
@@ -32,7 +32,7 @@ double *vectora;
 double *vectorb;
 double *vectorc;
 
-int batchsize;
+unsigned long batchsize;
 
 long long globalcounter = 0;
