+++ cudampilib/include/utility.h
@@ -8,7 +8,7 @@
 #include <sys/types.h>
 
 #ifdef ENABLE_OUTPUT_LOGS
-void save_vector_output_double(double* start, int batchsize, const char* filename, const char* header) {
+void save_vector_output_double(double* start, unsigned long batchsize, const char* filename, const char* header) {
     struct stat st = {0};
     if (stat("logs", &st) == -1) {
         mkdir("logs", 0700);
@@ -19,14 +19,14 @@ void save_vector_output_double(double* start, int batchsize, const char* filenam
 
     FILE *file = fopen(fullpath, "w");
     fprintf(file, "%s\n", header);
-     fprintf(file, "Array of size %d:\n", batchsize);
-    for (int i = 0; i < batchsize; i++) {
+     fprintf(file, "Array of size %ld:\n", batchsize);
+    for (unsigned long i = 0; i < batchsize; i++) {
         fprintf(file, "%f\n", start[i]);
     }
     fclose(file);
 }
 
-void save_vector_output_char(char* start, int batchsize, const char* filename, const char* header) {
+void save_vector_output_char(char* start, unsigned long batchsize, const char* filename, const char* header) {
     struct stat st = {0};
     if (stat("logs", &st) == -1) {
         mkdir("logs", 0700);
@@ -37,15 +37,15 @@ void save_vector_output_char(char* start, int batchsize, const char* filename, c
 
     FILE *file = fopen(fullpath, "w");
     fprintf(file, "%s\n", header);
-     fprintf(file, "Array of size %d:\n", batchsize);
-    for (int i = 0; i < batchsize; i++) {
+     fprintf(file, "Array of size %ld:\n", batchsize);
+    for (unsigned long i = 0; i < batchsize; i++) {
         fprintf(file, "%c\n", start[i]);
     }
     fclose(file);
 }
 #else
-void save_vector_output_double(double* start, int batchsize, const char* filename, const char* header) { }
-void save_vector_output_char(char* start, int batchsize, const char* filename, const char* header) { }
+void save_vector_output_double(double* start, unsigned long batchsize, const char* filename, const char* header) { }
+void save_vector_output_char(char* start, unsigned long batchsize, const char* filename, const char* header) { }
 #endif
 
 #endif // UTILITY_H
